backend/scripts/v4/trends/requirements.txt
playwright==1.49.1
beautifulsoup4==4.12.3
# psycopg2-binary version range — the strict 2.9.10 pin had no Apple Silicon
# wheel for Python 3.9, forcing a source build. The range below lets pip
# pick a binary wheel for whatever Python version the venv uses.
psycopg2-binary>=2.9.5,<3.0
